Update Projects Endpoint
Simplify how we save projects from API.

Instead of adding the projects and then making multiple dispatch calls to to add avatar images and workflows to the projects, we are adding the workflows and projects first and then saving the projects to the store.

Also, I added Actions for request, success and failure mainly to handle the ‘isLoading’ scenario. But they may be usefule in the future.