Redesign ProjectList and associated components
Instead of making all of the UI changes in the existing files, I rewrote them in new ones in their own directory.

I mainly did this because I found the existing implementation confusing. Especially with matter associated with the PopupMessage. It seemed like it’s animation was handles in multiple files.

My hope is that the new implementation is much more clear and straight forward.