Add a dynamic title option for the NavBar
Before this change, no pages had dynamic titles for the navbar.

Because of the way the nav bar is implemented with the library, we had no way to pass a title dynamically.

In order to allow for dynamic titles, I added actions to write to the store what title a page should have.

The nav bar still can except static text the same way it did before, so we only had to change the screens that will have dynamic titles.