Refactor how the classifier is stored in State
Before this change the classifier would be stored in state as a list of subjects, a current subject and an upcoming subject.

Because of this all of the action creators for the classifier were highly coupled with state.

With this refactor the new state structure is just a list of subjects and it’s up for the view to handle which subject should be displayed and when to load more.