Use a normal markdown library for Question
We were using a custom class called StyledMarkdown that made up for all of the custom markdown we have. But the Question text shouldn’t require any of the custom markdown we have.

This change improves the render time of the page. Before this change, the text would render after the rest of the view because it was being loaded into a webview