Add Drawing Screen
So far all of the functionality is to just show the drawing screen by itself. The use shouldn’t be able to do any classifying work.

ImageWithSvgOverlay - Just display an Image for now. In the future it will display an image with all of the shapes that have been drawn