Add Markable Image
This component is a combination of the subject image and the SVG overlay.

It will also manage all of the redux interactions for the SVG layer.

There is a lot of stuff commented out here. But I want to leave it because it will be uncommented in an upcoming pull request