Add New editing and deleting logic
It turns out that Svgs have trouble with touch events in modals in react-native.

So I had to rewrite how we tackle this problem. Basically, instead of adding pan responders to the svgs, the svgs report their location to ShapeEditorSvg and that component has a pan handling overlay that uses the shape’s locations to understand what the user wants to do when they touch and drag