Make sure refs exist before modifying
It is possible there is a race where `update` is called before the local refs to the svg shapes are set. Therefor we will check to see if they are defined before we access them.

This is a sporadic race condition and this solution will just appear as if the user missed what they are trying to tap.